refactor: merge request and response dto class
resolvse: #136
