+++ src/main/kotlin/com/kroffle/knitting/controller/handler/auth/dto/Authorized.kt
@@ -2,4 +2,8 @@ package com.kroffle.knitting.controller.handler.auth.dto
 
 import com.kroffle.knitting.controller.handler.helper.response.type.ObjectPayload
 
-class AuthorizedResponse(val token: String) : ObjectPayload
+object Authorized {
+    data class Response(
+        val token: String,
+    ) : ObjectPayload
+}
