+++ src/main/kotlin/com/kroffle/knitting/controller/handler/design/dto/MyDesign.kt
@@ -4,15 +4,17 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import com.kroffle.knitting.controller.handler.helper.response.type.ListItemPayload
 import java.time.OffsetDateTime
 
-class MyDesign(
-    val id: Long,
-    val name: String,
-    val yarn: String,
-    @JsonProperty("cover_image_url")
-    val coverImageUrl: String,
-    val tags: List<String>,
-    @JsonProperty("created_at")
-    val createdAt: OffsetDateTime,
-) : ListItemPayload {
-    override fun getCursor(): String = id.toString()
+object MyDesign {
+    data class Response(
+        val id: Long,
+        val name: String,
+        val yarn: String,
+        @JsonProperty("cover_image_url")
+        val coverImageUrl: String,
+        val tags: List<String>,
+        @JsonProperty("created_at")
+        val createdAt: OffsetDateTime,
+    ) : ListItemPayload {
+        override fun getCursor(): String = id.toString()
+    }
 }
