--- src/main/kotlin/com/kroffle/knitting/controller/handler/design/dto/NewDesignResponse.kt
@@ -1,7 +1,0 @@
-package com.kroffle.knitting.controller.handler.design.dto
-
-import com.kroffle.knitting.controller.handler.helper.response.type.ObjectPayload
-
-data class NewDesignResponse(
-    val id: Long,
-) : ObjectPayload
