--- src/main/kotlin/com/kroffle/knitting/controller/handler/knitter/dto/SalesSummaryResponse.kt
@@ -1,11 +1,0 @@
-package com.kroffle.knitting.controller.handler.knitter.dto
-
-import com.fasterxml.jackson.annotation.JsonProperty
-import com.kroffle.knitting.controller.handler.helper.response.type.ObjectPayload
-
-data class SalesSummaryResponse(
-    @JsonProperty("number_of_products_on_sales")
-    val numberOfProductsOnSales: Long,
-    @JsonProperty("number_of_products_sold")
-    val numberOfProductsSold: Long,
-) : ObjectPayload
