+++ src/main/kotlin/com/kroffle/knitting/controller/handler/product/ProductHandler.kt
@@ -5,14 +5,11 @@ import com.kroffle.knitting.controller.handler.helper.auth.AuthHelper
 import com.kroffle.knitting.controller.handler.helper.exception.ExceptionHelper
 import com.kroffle.knitting.controller.handler.helper.pagination.PaginationHelper
 import com.kroffle.knitting.controller.handler.helper.response.ResponseHelper
-import com.kroffle.knitting.controller.handler.product.dto.EditProductContentRequest
-import com.kroffle.knitting.controller.handler.product.dto.EditProductContentResponse
-import com.kroffle.knitting.controller.handler.product.dto.EditProductPackageRequest
-import com.kroffle.knitting.controller.handler.product.dto.EditProductPackageResponse
-import com.kroffle.knitting.controller.handler.product.dto.GetMyProductResponse
-import com.kroffle.knitting.controller.handler.product.dto.GetMyProductsResponse
-import com.kroffle.knitting.controller.handler.product.dto.RegisterProductRequest
-import com.kroffle.knitting.controller.handler.product.dto.RegisterProductResponse
+import com.kroffle.knitting.controller.handler.product.dto.EditProductContent
+import com.kroffle.knitting.controller.handler.product.dto.EditProductPackage
+import com.kroffle.knitting.controller.handler.product.dto.GetMyProduct
+import com.kroffle.knitting.controller.handler.product.dto.GetMyProducts
+import com.kroffle.knitting.controller.handler.product.dto.RegisterProduct
 import com.kroffle.knitting.domain.product.entity.Product
 import com.kroffle.knitting.domain.product.enum.ProductItemType
 import com.kroffle.knitting.domain.product.value.ProductItem
@@ -37,8 +34,8 @@ import java.util.stream.Collectors.toList
 class ProductHandler(private val productService: ProductService) {
     fun editProductPackage(req: ServerRequest): Mono<ServerResponse> {
         val knitterId = AuthHelper.getKnitterId(req)
-        val bodyMono: Mono<EditProductPackageRequest> = req
-            .bodyToMono(EditProductPackageRequest::class.java)
+        val bodyMono: Mono<EditProductPackage.Request> = req
+            .bodyToMono(EditProductPackage.Request::class.java)
             .switchIfEmpty(Mono.error(EmptyBodyException()))
 
         val product: Mono<Product> = bodyMono
@@ -64,15 +61,15 @@ class ProductHandler(private val productService: ProductService) {
             .flatMap {
                 ResponseHelper
                     .makeJsonResponse(
-                        EditProductPackageResponse(it.id!!)
+                        EditProductPackage.Response(it.id!!)
                     )
             }
     }
 
     fun editProductContent(req: ServerRequest): Mono<ServerResponse> {
         val knitterId = AuthHelper.getKnitterId(req)
-        val bodyMono: Mono<EditProductContentRequest> = req
-            .bodyToMono(EditProductContentRequest::class.java)
+        val bodyMono: Mono<EditProductContent.Request> = req
+            .bodyToMono(EditProductContent.Request::class.java)
             .switchIfEmpty(Mono.error(EmptyBodyException()))
 
         val product: Mono<Product> = bodyMono
@@ -91,15 +88,15 @@ class ProductHandler(private val productService: ProductService) {
             .flatMap {
                 ResponseHelper
                     .makeJsonResponse(
-                        EditProductContentResponse(it.id!!)
+                        EditProductContent.Response(it.id!!)
                     )
             }
     }
 
     fun registerProduct(req: ServerRequest): Mono<ServerResponse> {
         val knitterId = AuthHelper.getKnitterId(req)
-        val bodyMono: Mono<RegisterProductRequest> = req
-            .bodyToMono(RegisterProductRequest::class.java)
+        val bodyMono: Mono<RegisterProduct.Request> = req
+            .bodyToMono(RegisterProduct.Request::class.java)
             .switchIfEmpty(Mono.error(EmptyBodyException()))
 
         val product: Mono<Product> = bodyMono
@@ -116,7 +113,7 @@ class ProductHandler(private val productService: ProductService) {
             .flatMap {
                 ResponseHelper
                     .makeJsonResponse(
-                        RegisterProductResponse(it.id!!)
+                        RegisterProduct.Response(it.id!!)
                     )
             }
     }
@@ -136,7 +133,7 @@ class ProductHandler(private val productService: ProductService) {
             .flatMap {
                 ResponseHelper
                     .makeJsonResponse(
-                        GetMyProductResponse(
+                        GetMyProduct.Response(
                             id = it.id!!,
                             name = it.name,
                             fullPrice = it.fullPrice.value,
@@ -172,7 +169,7 @@ class ProductHandler(private val productService: ProductService) {
             .doOnError { ExceptionHelper.raiseException(it) }
             .map {
                 product ->
-                GetMyProductsResponse(
+                GetMyProducts.Response(
                     id = product.id!!,
                     name = product.name,
                     fullPrice = product.fullPrice.value,
