--- src/main/kotlin/com/kroffle/knitting/controller/handler/product/dto/EditProductContentRequest.kt
@@ -1,6 +1,0 @@
-package com.kroffle.knitting.controller.handler.product.dto
-
-data class EditProductContentRequest(
-    val id: Long,
-    val content: String,
-)
