--- src/main/kotlin/com/kroffle/knitting/controller/handler/product/dto/EditProductContentResponse.kt
@@ -1,7 +1,0 @@
-package com.kroffle.knitting.controller.handler.product.dto
-
-import com.kroffle.knitting.controller.handler.helper.response.type.ObjectPayload
-
-data class EditProductContentResponse(
-    val id: Long,
-) : ObjectPayload
