--- src/main/kotlin/com/kroffle/knitting/controller/handler/product/dto/EditProductPackageResponse.kt
@@ -1,7 +1,0 @@
-package com.kroffle.knitting.controller.handler.product.dto
-
-import com.kroffle.knitting.controller.handler.helper.response.type.ObjectPayload
-
-data class EditProductPackageResponse(
-    val id: Long,
-) : ObjectPayload
