--- src/main/kotlin/com/kroffle/knitting/controller/handler/product/dto/RegisterProductRequest.kt
@@ -1,5 +1,0 @@
-package com.kroffle.knitting.controller.handler.product.dto
-
-data class RegisterProductRequest(
-    val id: Long,
-)
