+++ src/test/kotlin/com/kroffle/knitting/helper/extension/MyDesign.kt
@@ -2,7 +2,7 @@ package com.kroffle.knitting.helper.extension
 
 import com.kroffle.knitting.controller.handler.design.dto.MyDesign
 
-fun MyDesign.like(other: MyDesign): Boolean {
+fun MyDesign.Response.like(other: MyDesign.Response): Boolean {
     if (this === other) return true
     return id == other.id &&
         name == other.name &&
